abstract repeated code for various devices to parent InterfaceDevice class
